pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ios-screen-time-screenshot-processing"
version = "0.1.0"
description = "Application for extracting battery and screen time usage data from iOS screenshots using OCR"
readme = "README.md"
authors = [
    {name = "iOS Screen Time Screenshot Processing Team"}
]
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    "httpx>=0.27.0",
    "matplotlib>=3.8.0",
    "numpy>=1.23",
    "opencv-python>=4.8.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    # Shared monorepo packages (installed by path in Docker)
    "global-pass-honor-username-auth>=0.1.0",
    "fastapi-errors>=0.1.0",
    "fastapi-logging>=0.1.0",
    "fastapi-ratelimit>=0.1.0",
    "db-toolkit>=0.1.0",
    "fastapi-pagination>=0.1.0",
    "fastapi-files>=0.1.0",
    "fastapi-tasks>=0.1.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "opencv-python-headless>=4.8.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "aiofiles>=24.1.0",
    # Note: slowapi replaced by fastapi-ratelimit from monorepo packages
]

gui = [
    "PyQt6>=6.9.0",
]

ocr = [
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
]

dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "basedpyright>=1.18.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
screenshot-gui = "screenshot_processor.gui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint.per-file-ignores]
# Scripts need sys.path manipulation before imports
"scripts/*" = ["E402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.basedpyright]
include = ["src", "tests"]
exclude = [
    "**/__pycache__",
    "**/node_modules",
    "**/.venv",
    "**/venv",
    "frontend",
    "alembic/versions",
]
typeCheckingMode = "standard"
pythonVersion = "3.11"
pythonPlatform = "All"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = "warning"
reportUnusedVariable = "warning"
reportDuplicateImport = "warning"
# Relax numpy/pandas type issues (common in scientific computing)
reportArgumentType = "warning"
reportOperatorIssue = "warning"
# Relax Optional handling (many valid runtime patterns)
reportOptionalSubscript = "warning"
reportOptionalMemberAccess = "warning"
reportOptionalOperand = "warning"
# Relax ORM/dynamic attribute access (SQLAlchemy, pandas, Celery)
reportAttributeAccessIssue = "warning"
reportCallIssue = "warning"
reportReturnType = "warning"
reportAssignmentType = "warning"
reportFunctionMemberAccess = "warning"  # Celery task.s() signatures

# NOTE: [tool.uv.sources] removed for Docker compatibility.
# In Docker, packages are installed by path from the monorepo.
# For local dev, use: uv sync (workspace root) or pip install -e
